Modularize zen
