+++ src/render/mod.rs
@@ -1,5 +1,4 @@
 // src/render/mod.rs
-pub mod zen;
 pub mod status;
 pub mod shortcuts_overlay;
 pub mod spotlight;
@@ -11,7 +10,7 @@ pub mod zoom_overlay;
 pub mod traits;
 pub mod settings;
 
-pub use crate::zen::render::render_zen;
+pub use crate::zen::view::render_zen;
 pub use status::render_status_bar;
 pub use shortcuts_overlay::render_shortcuts_overlay;
 pub use spotlight::render_spotlight;
