+++ src/state/core.rs
@@ -3,7 +3,7 @@ use std::time::Instant;
 use crate::node::{Node, NodeID, NodeMap};
 use crate::layout::{GEMX_HEADER_HEIGHT, LayoutRole};
 use crate::plugin::PluginHost;
-pub use crate::state::zen::*;
+pub use crate::zen::state::*;
 
 use super::hotkeys::load_hotkeys;
