+++ src/zen/mod.rs
@@ -1,3 +1,11 @@
-pub mod render;
+pub mod editor;
+pub mod state;
+pub mod view;
 pub mod journal;
 pub mod utils;
+
+pub use editor::*;
+pub use state::*;
+pub use view::*;
+pub use journal::*;
+pub use utils::*;
